refactor: moves SquareUI and dependant components to its own file
